requirements.txt
anyconfig>=0.9.9
oauthlib>=3.0.1
psycopg2>=2.8.2
python-dateutil>=2.8.0
PyYAML>=5.1.1
requests>=2.21.0
requests-oauthlib>=1.2.0
TwitterAPI>=2.5.9
urllib3>=1.24.3
beautifulsoup4==4.7.1
Flask==1.1.1
gensim==3.8.0
lxml==4.3.4
nltk==3.4.3
